apps/landing-page/apps/wozif-connect/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
python-dotenv
pydantic
asyncio
aiohttp
